\name{ET_functions-package}
\alias{ET_functions-package}
\alias{ET_functions}
\docType{package}
\title{
\packageTitle{ET_functions}
}
\description{
\packageDescription{ET_functions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ET_functions}
\packageIndices{ET_functions}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ET_functions}

Maintainer: \packageMaintainer{ET_functions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
