% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interception.R
\name{interception}
\alias{interception}
\title{Canopy and Stem Interception Model}
\usage{
calculate Interception for a day
}
\arguments{
\item{P.rate}{(mm/hr) is the mean precipitation rate.}

\item{Evap}{(mm) total evporation (Here we try to aplly the Penman-montieth ET.)}

\item{P}{(mm) total precipitation}

\item{c}{(-) canopy cover per unit area}

\item{S}{(mm) canopy storage capacity}

\item{St}{(mm) trunks storage capacity}

\item{Pt}{(-) the propotion of rain goes to stemflow}
}
\value{
P.canopy (mm) canopy interception, P.stem (mm) stem interception, tot_inter (mm) total interception, and prc.inter (%) precentage of interception in precipitation. .
}
\description{
This model calculates the precipitation necessery to fill canopy and stem storage, based on Gash et al. 1995, and Valente et al. 1977.
The defults represent values of Mediterranean Quercus ilex (evergreen oak) ecosystem acording to the reference below.
}
\author{
Elad Dente
}

