% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penman_montieth_ed.R
\name{penman_montieth}
\alias{penman_montieth}
\title{Penman Montieth Equation}
\usage{
penman_montieth(Tair, vpd, Rnet, gs, ga, dayl = 11, CP = 1010,
  Pair = 101325)
}
\arguments{
\item{Tair}{(deg C) air temperature}

\item{vpd}{(Pa)    vapour pressure deficit}

\item{Rnet}{(j/m2/day)      net radiation}

\item{gs}{(mm/s)  surface conductance}

\item{ga}{(mm/s)  aerodynamic conductance}

\item{CP}{=      1010.0 (J/kg*K) specific heat of air}

\item{Pair}{= 101325 (Pa)}
}
\value{
ET (W/m^2),Evapotranspiration (m/s), Evapotranspiration (mm/day)
}
\description{
THis function computer evapotranspiration based on radiation, conductance etc
}
\author{
Naomi (edited by Elad Dente)
}

