% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCS.R
\name{SCS}
\alias{SCS}
\title{SCS Model}
\usage{
SCS(P, I, S)
}
\arguments{
\item{P}{(mm) daily rain.}

\item{I}{(mm) daily initial abstraction (we try to use the daily interception from the interception model)}

\item{S}{(-)  (from the CN number. default is 130 for CN 66)}
}
\value{
Q (mm) runoff
}
\description{
This function calculate the SCS method for Runoff. This function atempt to use physical parameters rather the curve number of the original method.
Here the usage is for estimation of daily runoff
}
\author{
Elad Dente
}
\references{
http://www.professorpatel.com/uploads/7/6/5/6/7656897/tr55.pdf
}

