% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_conductance.R
\name{canopy_conductance}
\alias{canopy_conductance}
\title{Canopy Conductance Model}
\usage{
calculate canopy conductance, based on the Jarvis model. To be used in Penman-Montieth ET model.
}
\arguments{
\item{Cleaf_max}{(mm/s) maximum stomatal contuctance .}

\item{Rnrad}{(-) Relative solar radiation.}

\item{Rvpd}{(-) Relative absolute-humidity deficit.}

\item{T}{(deg c) air temperature at measured height above ground (0>T<40 deg C).}

\item{Rteta}{(-) Relative soli moisture deficit.}

\item{LAI}{(-) Leaf area index.}

\item{Fs}{(-) Shelter factor.}
}
\value{
Jarvis model coefficient (Jarvis_c) and canopy conductance (Kc, mm/s)
}
\description{
The R in Jarvis parameters is for "relative". the valus can be taken from the plots and equations in "Physical Hydrology: Third Edition" by S. Lawrence Dingman (p. 282).
}
\author{
Elad Dente
}

