% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCS.R
\name{SCS}
\alias{SCS}
\title{SCS Model}
\usage{
SCS(P, I)
}
\arguments{
\item{P}{(mm) daily rain.}

\item{I}{(mm) daily initial abstraction (we try to use the daily interception from the interception model)}
}
\value{
I, S, CN and Q (all in mm)
}
\description{
This function calculate the SCS method for Runoff, in atempt to use physical parameters rather the curve number of #' the original method.
Here the usage is for estimation of daily runoff
}
\author{
Elad Dente
}
\references{
http://www.professorpatel.com/uploads/7/6/5/6/7656897/tr55.pdf
}

